==> Constructing_ML_Model/1_Creating_URL.ipynb <==
# Data downloading and creating Comma Seperated Value(.csv) Files

## Selecting and Downloading 5000 of Phishing URLs and 5000 of Bnigan URLs


*   In here, we collect Phishing URLs from http://data.phishtank.com/data/online-valid.csv and  select randomly 5000 URLs and save in to phishing.csv file
*   Also, we collect Benigan URLs from https://www.unb.ca/cic/datasets/url-2016.html and  select randomly 5000 URLs from Benign_list_big_final.csv file  and save in to Benigan.csv file



#importing required packages for this Notebook
import pandas as pd

### 1. Selecting and Creating Phishing URLs

data0 = pd.read_csv("CSV Files/1-valid_Phishing.csv")
data0.head() 

data0.shape

phishurl = data0.sample(n = 5000, random_state = 12).copy()
phishurl = phishurl["url"]
phishurl = phishurl.reset_index(drop=True)
phishurl.head()

# Storing the extracted legitimate URLs fatures to csv file
phishurl.to_csv('CSV Files/3-phishurl.csv', index= False)

### 1. Selecting and Creating Benigan URLs

data1 = pd.read_csv("CSV Files/2-valid_Benign.csv")
data1.columns = ['URLs']
data1.head()

#Collecting 5,000 Legitimate URLs randomly
legiurl = data1.sample(n = 5000, random_state = 12).copy()
legiurl = legiurl.reset_index(drop=True)
legiurl.shape

# Storing the extracted legitimate URLs fatures to csv file
legiurl.to_csv('CSV Files/4-beniganurl.csv', index= False)

### Contributors - Deshitha Hansajith [Linked In](https://www.linkedin.com/in/deshitha-hansajith/), Nipuni Dilukshika [Linked In](https://www.linkedin.com/in/nipuni-dilukshika-186764197/)
